Split out to utils.py file
